# docs/api/openapi/openapi.yaml
...
{
  "openapi": "3.1.0",
  "info": {
    "title": "家庭认知健康日记 API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/consents": {
      "get": {
        "tags": [
          "consents"
        ],
        "summary": "List Consents",
        "operationId": "list_consents_v1_consents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "consents"
        ],
        "summary": "Grant Consent",
        "operationId": "grant_consent_v1_consents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/consents/{consentType}": {
      "delete": {
        "tags": [
          "consents"
        ],
        "summary": "Revoke Consent",
        "operationId": "revoke_consent_v1_consents__consentType__delete",
        "parameters": [
          {
            "name": "consentType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Consenttype"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentIn": {
        "properties": {
          "type": {
            "type": "string",
            "pattern": "^(SENSITIVE|ELDER_INFORMED|RESEARCH|PRODUCT)$",
            "title": "Type"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "type",
          "version"
        ],
        "title": "ConsentIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}